Wire in rabin to do 'real' chunk splitting...
'real' in the sense that we only test with tiny example data right
now.

+++ src/hash.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef unsigned long hash;
+typedef unsigned hash;
